Python/requirements.txt
Pillow==10.1.0
opencv-python==4.8.1.78
PySide6==6.6.1
numpy==1.24.3
